# perf(api): improve multi payments queries