--- a/packages/api-database/source/migrations/1697617471901-CreateIndexes.ts
+++ b/packages/api-database/source/migrations/1697617471901-CreateIndexes.ts
@@ -5,12 +5,12 @@ export class CreateIndexes1697617471901 implements MigrationInterface {
 		// language=postgresql
 		await queryRunner.query(`
             CREATE UNIQUE INDEX transactions_sender_nonce ON transactions(sender_public_key, nonce);
-            CREATE INDEX transactions_to_block_number_tx_index ON transactions("to", block_number DESC, transaction_index DESC);
-            CREATE INDEX transactions_from_block_number_tx_index ON transactions("from", block_number DESC, transaction_index DESC);
-            CREATE INDEX transactions_sender_block_number_tx_index ON transactions(sender_public_key, block_number DESC, transaction_index DESC);
+            CREATE INDEX transactions_to_timestamp_tx_index ON transactions("to", timestamp DESC, transaction_index DESC);
+            CREATE INDEX transactions_from_timestamp_tx_index ON transactions("from", timestamp DESC, transaction_index DESC);
+            CREATE INDEX transactions_sender_timestamp_tx_index ON transactions(sender_public_key, timestamp DESC, transaction_index DESC);
+            CREATE INDEX transactions_multi_payment_recipients ON transactions USING GIN (multi_payment_recipients);
 
             CREATE INDEX transactions_block_hash ON transactions(block_hash);
-            CREATE INDEX transactions_block_number_transaction_index ON transactions(block_number, transaction_index);
 
             CREATE INDEX transactions_value ON transactions(value);
             CREATE INDEX transactions_gas_price ON transactions(gas_price);
@@ -40,6 +40,7 @@ export class CreateIndexes1697617471901 implements MigrationInterface {
             CREATE INDEX blocks_validator_round ON blocks(validator_round);
 
             CREATE INDEX multi_payments_hash_to ON multi_payments(hash, "to") WHERE success IS TRUE;
+            CREATE INDEX multi_payments_to ON multi_payments("to") WHERE success IS TRUE;
 
             CREATE INDEX wallets_balance ON wallets(balance);
             CREATE INDEX wallets_attributes ON wallets using GIN(attributes);
@@ -57,12 +58,12 @@ export class CreateIndexes1697617471901 implements MigrationInterface {
 		// language=postgresql
 		await queryRunner.query(`
             DROP INDEX transactions_sender_nonce;
-            DROP INDEX transactions_to_block_number_tx_index;
-            DROP INDEX transactions_sender_block_number_tx_index;
-            DROP INDEX transactions_from_block_number_tx_index;
+            DROP INDEX transactions_to_timestamp_tx_index;
+            DROP INDEX transactions_sender_timestamp_tx_index;
+            DROP INDEX transactions_from_timestamp_tx_index;
+            DROP INDEX transactions_multi_payment_recipients;
 
             DROP INDEX transactions_block_hash;
-            DROP INDEX transactions_block_number_transaction_index;
 
             DROP INDEX transactions_value;
             DROP INDEX transactions_gas_price;
@@ -79,6 +80,9 @@ export class CreateIndexes1697617471901 implements MigrationInterface {
             DROP INDEX transactions_timestamp_asc_transaction_index_desc;
             DROP INDEX transactions_function_sig_address;
             DROP INDEX transactions_deployed_contract_index;
+
+            DROP INDEX multi_payments_hash_to;
+            DROP INDEX multi_payments_to;
 
             DROP INDEX blocks_transactions_count;
             DROP INDEX blocks_reward;