--- a/packages/api-database/source/search/filters/transaction-filter.ts
+++ b/packages/api-database/source/search/filters/transaction-filter.ts
@@ -163,8 +163,7 @@ export class TransactionFilter {
 
 		const multipaymentRecipientExpression: MultiPaymentExpression = {
 			op: "multiPayment",
-			property: "to",
-			value: criteria,
+			value: [criteria],
 		};
 
 		return {