--- a/packages/api-database/source/search/types/expressions.ts
+++ b/packages/api-database/source/search/types/expressions.ts
@@ -79,9 +79,8 @@ export type FunctionSigExpression<TEntity> = {
 };
 
 export type MultiPaymentExpression = {
-	property: "from" | "to";
 	op: "multiPayment";
-	value: string;
+	value: string[];
 };
 
 export type NotNullExpression<TEntity> = {