--- a/packages/api-http/source/controllers/receipts.ts
+++ b/packages/api-http/source/controllers/receipts.ts
@@ -88,7 +88,7 @@ export class ReceiptsController extends Controller {
 		return [
 			{
 				direction: "desc",
-				property: "blockNumber",
+				property: "timestamp",
 			},
 			{
 				direction: "desc",