--- a/packages/api-sync/source/service.ts
+++ b/packages/api-sync/source/service.ts
@@ -183,10 +183,12 @@ export class Sync implements Contracts.ApiSync.Service {
 
 				logs: receipt.logs,
 
+				multiPaymentRecipients:
+					parsedMultiPayments.length > 0 ? [...new Set(parsedMultiPayments.map((mp) => mp.to))] : undefined,
+
 				nonce: data.nonce.toFixed(),
 
 				output: receipt.output,
-
 				senderPublicKey: data.senderPublicKey,
 				signature: formatEcdsaSignature(data.r!, data.s!, data.v!),
 				status: receipt.status,
@@ -194,9 +196,6 @@ export class Sync implements Contracts.ApiSync.Service {
 				to: data.to,
 				transactionIndex: data.transactionIndex!,
 				value: data.value.toFixed(),
-				multiPaymentRecipients: parsedMultiPayments.length
-					? [...new Set(parsedMultiPayments.map((mp) => mp.to))]
-					: undefined,
 			});
 
 			multiPayments.push(...parsedMultiPayments);