<!--
Thanks for your interest in the project. Bugs filed and PRs submitted are appreciated!

Please make sure you're familiar with and follow the instructions in the [contributing guidelines](https://ark.dev/docs/program-incentives/guidelines/contributing).

Please fill out the information below to expedite the review and (hopefully) merge of your pull request!
-->

## Summary

<!-- What changes are being made? -->

`/transactions?to` and related endpoints hit the `multi_payments` table just to check for existence of a recipient and are currently pretty slow. Adding a dedicated `multi_payment_recipients` column to the `transactions` table to store all multi payment related recipients helps quite a bit here.

<details>
<summary>Before</summary>

```
Limit  (cost=0.42..15947.05 rows=100 width=1224) (actual time=246.868..438.452 rows=100 loops=1)
  ->  Index Scan Backward using transactions_timestamp_transaction_index on transactions "Transaction"  (cost=0.42..45910804.02 rows=287903 width=1224) (actual time=246.866..438.429 rows=100 loops=1)
        Filter: (("to" = '0xaF076328c6b792B7bE142C87038b4C4B2711f686'::citext) OR (hashed SubPlan 2))
        Rows Removed by Filter: 169091
        SubPlan 2
          ->  Gather  (cost=1000.00..33538.41 rows=324 width=65) (actual time=0.482..245.572 rows=348 loops=1)
                Workers Planned: 2
                Workers Launched: 2
                ->  Parallel Seq Scan on multi_payments mp  (cost=0.00..32506.01 rows=135 width=65) (actual time=0.619..232.522 rows=116 loops=3)
                      Filter: (success AND ("to" = '0xaF076328c6b792B7bE142C87038b4C4B2711f686'::citext))
                      Rows Removed by Filter: 463117
Planning Time: 0.309 ms
Execution Time: 438.500 ms
```

</details>


<!-- What changes are being made? -->

<details>
<summary>After</summary>

```
Limit  (cost=1052.93..1053.18 rows=100 width=1224) (actual time=0.486..0.493 rows=100 loops=1)
  ->  Sort  (cost=1052.93..1053.59 rows=265 width=1224) (actual time=0.485..0.488 rows=100 loops=1)
        Sort Key: "timestamp" DESC, transaction_index DESC
        Sort Method: top-N heapsort  Memory: 428kB
        ->  Bitmap Heap Scan on transactions "Transaction"  (cost=27.41..1042.80 rows=265 width=1224) (actual time=0.089..0.345 rows=348 loops=1)
              Recheck Cond: (("to" = '0xaF076328c6b792B7bE142C87038b4C4B2711f686'::citext) OR (multi_payment_recipients @> '{0xaF076328c6b792B7bE142C87038b4C4B2711f686}'::citext[]))
              Heap Blocks: exact=348
              ->  BitmapOr  (cost=27.41..27.41 rows=265 width=0) (actual time=0.063..0.064 rows=0 loops=1)
                    ->  Bitmap Index Scan on transactions_to_timestamp_tx_index  (cost=0.00..4.67 rows=16 width=0) (actual time=0.021..0.021 rows=0 loops=1)
                          Index Cond: ("to" = '0xaF076328c6b792B7bE142C87038b4C4B2711f686'::citext)
                    ->  Bitmap Index Scan on transactions_multi_payment_recipients  (cost=0.00..22.61 rows=250 width=0) (actual time=0.041..0.041 rows=348 loops=1)
                          Index Cond: (multi_payment_recipients @> '{0xaF076328c6b792B7bE142C87038b4C4B2711f686}'::citext[])
Planning Time: 0.149 ms
Execution Time: 0.516 ms
```

</details>

<!-- Why are these changes necessary? -->

<!-- How were these changes implemented? -->

## Checklist

<!-- Have you done all of these things?  -->

- [ ] Documentation _(if necessary)_
- [ ] Tests _(if necessary)_
- [ ] Ready to be merged

<!-- Feel free to add additional comments. -->